+++ akshare/__init__.py
@@ -2323,9 +2323,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.5 fix: fix stock_info_a_code_name interface
 1.9.6 fix: fix futures_hog_info interface
 1.9.7 add: add stock_profit_forecast_ths interface
+1.9.8 fix: fix stock_hk_valuation_baidu interface
 """
 
-__version__ = "1.9.7"
+__version__ = "1.9.8"
 __author__ = "AKFamily"
 
 import sys