+++ docs/changelog.md
@@ -50,6 +50,10 @@
 
 ## 更新说明
 
+1.9.8 fix: fix stock_hk_valuation_baidu interface
+
+    1. 修复 stock_hk_valuation_baidu 接口
+
 1.9.7 add: add stock_profit_forecast_ths interface
 
     1. 新增 stock_profit_forecast_ths 接口
@@ -1977,6 +1981,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.8 fix: fix stock_hk_valuation_baidu interface
 
 1.9.7 add: add stock_profit_forecast_ths interface